:sparkles: feat: Validações referentes à autenticação na entidade Paciente

+++ src/main/java/com/connectCare/connectCareApi/services/impl/PacienteServiceImpl.java
@@ -1,5 +1,6 @@
 package com.connectCare.connectCareApi.services.impl;
 
+import com.connectCare.connectCareApi.exceptions.NaoAutorizadoException;
 import com.connectCare.connectCareApi.exceptions.NenhumRegistroEncontradoException;
 import com.connectCare.connectCareApi.exceptions.OperacaoBancoDeDadosException;
 import com.connectCare.connectCareApi.exceptions.PacienteNaoEncontradoException;
@@ -7,12 +8,15 @@ import com.connectCare.connectCareApi.models.entities.Paciente;
 import com.connectCare.connectCareApi.models.entities.Usuario;
 import com.connectCare.connectCareApi.repositories.PacienteRepository;
 import com.connectCare.connectCareApi.services.GenericService;
+import com.connectCare.connectCareApi.utils.UsuarioAutenticado;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class PacienteServiceImpl implements GenericService<Paciente> {
@@ -26,6 +30,12 @@ public class PacienteServiceImpl implements GenericService<Paciente> {
     @Override
     public Paciente create(Paciente paciente) {
     	try {
+
+            //Verificando se o ID do usuário autenticado é igual ao ID do usuário associado ao paciente
+            if(!Objects.equals(paciente.getUsuario().getId(), UsuarioAutenticado.getUsuarioAutenticado().getId())){
+                throw new NaoAutorizadoException();
+            }
+
     		Usuario usuarioEncontrado = usuarioService.getById(paciente.getUsuario().getId());
             paciente.setUsuario(usuarioEncontrado);
             return repository.save(paciente);
@@ -37,11 +47,22 @@ public class PacienteServiceImpl implements GenericService<Paciente> {
 
     @Override
     public Paciente getById(Integer id) {
-        return repository.findById(id).orElseThrow(() -> new PacienteNaoEncontradoException(id));
+        Paciente pacienteEncontrado = repository.findById(id).orElseThrow(() -> new PacienteNaoEncontradoException(id));
+
+        //Verificando se o ID do usuário autenticado é igual ao ID do usuário associado ao paciente
+        if(!Objects.equals(pacienteEncontrado.getUsuario().getId(), UsuarioAutenticado.getUsuarioAutenticado().getId())){
+            throw new NaoAutorizadoException();
+        }
+
+        return pacienteEncontrado;
     }
 
     @Override
     public List<Paciente> getAll() {
+        //Verificando se o atual usuário é ADMIN
+        if(!UsuarioAutenticado.getUsuarioAutenticado().getAuthorities().contains(new SimpleGrantedAuthority("ADMIN"))){
+            throw new NaoAutorizadoException();
+        }
         List<Paciente> pacientesEncontrados = repository.findAll();
         if(pacientesEncontrados.isEmpty()) throw new NenhumRegistroEncontradoException("Paciente");
         return pacientesEncontrados;
@@ -51,6 +72,11 @@ public class PacienteServiceImpl implements GenericService<Paciente> {
     public Paciente update(Paciente paciente) {
         try{
             Paciente pacienteEncontrado = repository.getReferenceById(paciente.getId());
+
+            //Verificando se o ID do usuário autenticado é igual ao ID do usuário associado ao paciente
+            if(!Objects.equals(pacienteEncontrado.getUsuario().getId(), UsuarioAutenticado.getUsuarioAutenticado().getId())){
+                throw new NaoAutorizadoException();
+            }
 
             pacienteEncontrado.setCpf(paciente.getCpf());
             pacienteEncontrado.setEndereco(paciente.getEndereco());
@@ -73,6 +99,12 @@ public class PacienteServiceImpl implements GenericService<Paciente> {
     public void delete(Integer id) {
         try{
             Paciente pacienteEncontrado = repository.findById(id).orElseThrow(() -> new PacienteNaoEncontradoException(id));
+
+            //Verificando se o ID do usuário autenticado é igual ao ID do usuário associado ao paciente
+            if(!Objects.equals(pacienteEncontrado.getUsuario().getId(), UsuarioAutenticado.getUsuarioAutenticado().getId())){
+                throw new NaoAutorizadoException();
+            }
+
             repository.delete(pacienteEncontrado);
         }catch(DataIntegrityViolationException e) {
             throw new OperacaoBancoDeDadosException("Não foi possível excluir, pois esse paciente está relacionado com algum dependente.");
